Failed Partial App. and SpecialIndex for Controls
Ok, to avoid duplicate code in `createNamedControl*` I wanted to implement to partial application using the `bind` method. However, due to this partial function being used to modify the `String` prototype, I was having some issues with `this`. I couldn't figure it out, so I went back to the code duplication. I left the better code in for now--just in case anyone wants to fix it.

I also realized that I need to set the specialIndex of Control UGens to their respective index in the control array.